+++ UFO Tracker/Models/LocationSightings.cs
@@ -5,10 +5,14 @@
         public int Id { get; set; }
         public string Description { get; set; }
         public string DateOfEvent { get; set;}
+        public string Duration { get; set; }
+        public string Shape { get; set; }
         public string City { get; set; }
         public string State { get; set; }
         public string StreetAddress { get; set; }
         public int Zipcode { get; set; }
+        public float CityLongitude { get; set; }
+        public float CityLatitude { get; set; }
 
     }
 }
